requirements.txt
aiofiles==0.7.0
anyio==3.3.4
ariadne==0.13.0
asgiref==3.4.1
beautifulsoup4==4.10.0
boto3==1.18.63
botocore==1.21.63
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
fastapi-utils==0.2.1
graphql-core==3.1.6
greenlet==1.1.2
h11==0.12.0
idna==3.3
jmespath==0.10.0
numpy==1.21.2
pandas==1.3.4
psycopg2==2.9.1
pyarrow==5.0.0
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wincertstore==0.2
